Added constraint solver framework

+++ ParticleToy/Particle.h
@@ -17,6 +17,6 @@ public:
     Vec2f m_Position;
     Vec2f m_Velocity;
     Vec2f m_Force;
-    double m_Mass;
+    float m_Mass;
 
 };
